+++ public/styles.css
@@ -83,8 +83,8 @@
 }
 .modal-right-el {
   background-color: white;
-  border-bottom: 2px;
-  border-color: #767676;
+  border-bottom: 1px solid;
+  border-color: rgb(229, 229, 229);
   height: auto;
   display: block;
 }
@@ -145,7 +145,7 @@
 .required {
   font-family: Arial, Helvetica, sans-serif;
   padding: 0px 20px 0px 20px;
-  margin: 0;
+  margin: 0 0 11px 0;
   line-height: 19.5px;
   font-weight: 400px;
   font-size: 11.7px;
