+++ test/client/components/modal/UserReviewTitle.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
-import UserProductRating from '../../../../client/components/modal/UserProductRating';
+import UserReviewTitle from '../../../../client/components/modal/UserReviewTitle';
 
-describe('<UserProductRating />', () => {
+describe('<UserReviewTitle />', () => {
   test('Should have 2 children', () => {
     const expected = 2;
-    const wrapper = shallow(<UserProductRating />);
+    const wrapper = shallow(<UserReviewTitle />);
     expect(wrapper.find('.modal-right-el').children().length).toBe(expected);
   });
 });
